==> docs/00_project_setup.rmd <==
---
title: "00 Project Setup"
output: html_notebook
editor_options: 
  chunk_output_type: inline
---
<style>
    body .main-container {
        max-width: 100%;
    }
</style>


This file described dependencies and potential troubleshooting tips.

This paper was run using v1.0.0 of the MeasuringLandscape package. This release reflects the state of the code as of publication. It is the best chance at obtaining identical results as shown in print. Future releases will improve the presentation of the R-Notebooks and refactor the code to be more readable and reusable and may not generate exactly identical figures or messages. This version is archived so that even as the package improves over time, the analysis can be rerun against that specific version of the code.

```{r}
# install.packages("devtools") 
#devtools::install_github("rexdouglass/MeasuringLandscape") #@v1.0.0
```

You may hit an error "`maximal number of DLLs reached..." This is set by an environmental variable in R. We've specifcied it in the included ".Renviron" file "R_MAX_NUM_DLLS = 614"" which should work unless your system has a hard fixed lower level or you specifcy it elsehwere.


This package is highly dependent on specific package versions. Many packages updated their interfaces just over the course of writing this paper. We use a system called Packrat which records specific version numbers we used and installs them when you install this package.

```{r}

install.packages('packrat', dependencies=T)
library(packrat)
# packrat::init(project = paste0(here::here(), "/packrat"))

```

You should not have to manually install these packages, Packrat should take care of them. If for any reason that fails, these install commands may help you find the correct version of each package.


In order to use specific fonts in the pdf, we use a system called Cairo. To use Cairo, must also install from https://www.cairographics.org/download/
```{r}
install.packages('Cairo', dependencies=T)
```


```{r}

install.packages('devtools', dependencies=T)
devtools::install_github("dselivanov/LSHR") 
devtools::install_github("tidyverse/ggplot2") # geom_sf requires ggplot installed off of the dev server

install.packages('knitr', dependencies=T)
install.packages('tidyverse', dependencies=T)
install.packages('data.table', dependencies=T)
install.packages('sf', dependencies=T)
install.packages('caTools', dependencies=T)
install.packages('xgboost', dependencies=T)
install.packages('pacman', dependencies=T)
install.packages('units', dependencies=T)
install.packages('date', dependencies=T)
install.packages('car', dependencies=T)
install.packages('strigr', dependencies=T)
install.packages('cowplot', dependencies=T)
install.packages('ggjoy', dependencies=T)
install.packages('party', dependencies=T)
install.packages('Metrics', dependencies=T)
install.packages('MLmetrics', dependencies=T)
install.packages('precrec', dependencies=T)
install.packages('dummies', dependencies=T)
install.packages('janitor', dependencies=T)
install.packages('DT', dependencies=T)
install.packages('ggmap', dependencies=T)
install.packages('mosaic', dependencies=T)
install.packages('ggformula', dependencies=T)
install.packages('magrittr', dependencies=T)
install.packages('glue', dependencies=T)
install.packages('RANN', dependencies=T)
install.packages('re2r', dependencies=T)
install.packages('dummy', dependencies=T)
install.packages('doParallel', dependencies=T)
install.packages('quanteda', dependencies=T)
install.packages('ggrepel', dependencies=T)
install.packages('dissUtils', dependencies=T)
install.packages('digest', dependencies=T)
install.packages('stringi', dependencies=T)
install.packages('xtable', dependencies=T)
install.packages('SnowballC', dependencies=T)
install.packages('xtable', dependencies=T)
install.packages('tools', dependencies=T)
install.packages('lubridate', dependencies=T)
install.packages('digest', dependencies=T)
install.packages('rgdal', dependencies=T)
install.packages('stringr', dependencies=T)
install.packages('foreach', dependencies=T)
install.packages('pinp', dependencies=T)
install.packages('partykit', dependencies=T)
install.packages('raster', dependencies=T)


packrat::snapshot()

# must download 

devtools::document()

#library("miniCRAN")
#tags <- "MeasuringLandscape"
#pkgDep(tags, availPkgs = cranJuly2014)

require(devtools)
#install_version("backports", version = "1.1.0")
#library(backports)

```
